+++ taxi/projects.py
@@ -1,19 +1,12 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
 from collections import defaultdict
 import copy
 import datetime
 import json
 import os
-import re
-
-import six
 
 from .exceptions import TaxiException
 
 
-@six.python_2_unicode_compatible
 class Project:
     STATUS_NOT_STARTED = 0
     STATUS_ACTIVE = 1
@@ -34,7 +27,7 @@ class Project:
         STATUS_CANCELLED: 'C',
     }
 
-    def __init__(self, id, name, status=None, description=None, budget=None):
+    def __init__(self, id, name, status=None, description=None, budget=None, team=None):
         self.id = int(id)
         self.name = name
         self.activities = []
@@ -45,6 +38,7 @@ class Project:
         self.start_date = None
         self.end_date = None
         self.backend = None
+        self.team = team
 
     def __str__(self):
         if self.status in self.STATUSES:
@@ -119,7 +113,7 @@ Description: %s""" % (self.id, self.name, status, start_date, end_date,
         Converts a (project, activity) tuple to a string in the format xxx/yyy, or a (project, activity, role) tuple in
         the format xxx/yyy/zzz.
         """
-        return '/'.join(six.text_type(part) for part in t if part is not None)
+        return '/'.join(str(part) for part in t if part is not None)
 
 
 class Activity: